Move video_core/color.h to common/color.h
