+++ src/citra_qt/debugger/graphics_framebuffer.cpp
@@ -9,10 +9,11 @@
 #include <QPushButton>
 #include <QSpinBox>
 
+#include "common/color.h"
+
 #include "core/hw/gpu.h"
 #include "core/memory.h"
 
-#include "video_core/color.h"
 #include "video_core/pica.h"
 #include "video_core/utils.h"
