+++ src/common/CMakeLists.txt
@@ -24,6 +24,7 @@ set(HEADERS
             bit_field.h
             break_points.h
             chunk_file.h
+            color.h
             common_funcs.h
             common_paths.h
             common_types.h
